chore: resolve all remaining circular dependencies
