+++ index.js
@@ -27,7 +27,7 @@ exports.Pool = Pool;
 
 exports.PoolCluster = PoolCluster;
 
-exports.createServer = function(handler) {
+exports.createServer = function (handler) {
   const Server = require('./lib/server.js');
   const s = new Server();
   if (handler) {
@@ -45,33 +45,33 @@ exports.raw = SqlString.raw;
 
 exports.__defineGetter__(
   'createConnectionPromise',
-  () => require('./promise.js').createConnection
+  () => require('./promise.js').createConnection,
 );
 
 exports.__defineGetter__(
   'createPoolPromise',
-  () => require('./promise.js').createPool
+  () => require('./promise.js').createPool,
 );
 
 exports.__defineGetter__(
   'createPoolClusterPromise',
-  () => require('./promise.js').createPoolCluster
+  () => require('./promise.js').createPoolCluster,
 );
 
 exports.__defineGetter__('Types', () => require('./lib/constants/types.js'));
 
 exports.__defineGetter__('Charsets', () =>
-  require('./lib/constants/charsets.js')
+  require('./lib/constants/charsets.js'),
 );
 
 exports.__defineGetter__('CharsetToEncoding', () =>
-  require('./lib/constants/charset_encodings.js')
+  require('./lib/constants/charset_encodings.js'),
 );
 
-exports.setMaxParserCache = function(max) {
+exports.setMaxParserCache = function (max) {
   parserCache.setMaxCache(max);
 };
 
-exports.clearParserCache = function() {
+exports.clearParserCache = function () {
   parserCache.clearCache();
 };
