+++ lib/create_pool.js
@@ -7,4 +7,4 @@ function createPool(config) {
   return new Pool({ config: new PoolConfig(config) });
 }
 
-module.exports = createPool
+module.exports = createPool;
