+++ lib/inherit_events.js
@@ -3,7 +3,7 @@
 function inheritEvents(source, target, events) {
   const listeners = {};
   target
-    .on('newListener', eventName => {
+    .on('newListener', (eventName) => {
       if (events.indexOf(eventName) >= 0 && !target.listenerCount(eventName)) {
         source.on(
           eventName,
@@ -12,11 +12,11 @@ function inheritEvents(source, target, events) {
             args.unshift(eventName);
 
             target.emit.apply(target, args);
-          })
+          }),
         );
       }
     })
-    .on('removeListener', eventName => {
+    .on('removeListener', (eventName) => {
       if (events.indexOf(eventName) >= 0 && !target.listenerCount(eventName)) {
         source.removeListener(eventName, listeners[eventName]);
         delete listeners[eventName];
