+++ lib/promise_connection.js
@@ -1,10 +1,10 @@
 'use strict';
 
+const EventEmitter = require('events').EventEmitter;
 const PromisePreparedStatementInfo = require('./promise_prepared_statement_info.js');
 const makeDoneCb = require('./make_done_cb.js');
 const inheritEvents = require('./inherit_events.js');
-const { Connection } = require('../index.js');
-const EventEmitter = require('events').EventEmitter;
+const BaseConnection = require('./base_connection.js');
 
 class PromiseConnection extends EventEmitter {
   constructor(connection, promiseImpl) {
@@ -16,7 +16,7 @@ class PromiseConnection extends EventEmitter {
       'drain',
       'connect',
       'end',
-      'enqueue'
+      'enqueue',
     ]);
   }
 
@@ -29,7 +29,7 @@ class PromiseConnection extends EventEmitter {
     const localErr = new Error();
     if (typeof params === 'function') {
       throw new Error(
-        'Callback function is not available with promise clients.'
+        'Callback function is not available with promise clients.',
       );
     }
     return new this.Promise((resolve, reject) => {
@@ -47,7 +47,7 @@ class PromiseConnection extends EventEmitter {
     const localErr = new Error();
     if (typeof params === 'function') {
       throw new Error(
-        'Callback function is not available with promise clients.'
+        'Callback function is not available with promise clients.',
       );
     }
     return new this.Promise((resolve, reject) => {
@@ -61,7 +61,7 @@ class PromiseConnection extends EventEmitter {
   }
 
   end() {
-    return new this.Promise(resolve => {
+    return new this.Promise((resolve) => {
       this.connection.end(resolve);
     });
   }
@@ -97,7 +97,7 @@ class PromiseConnection extends EventEmitter {
     const c = this.connection;
     const localErr = new Error();
     return new this.Promise((resolve, reject) => {
-      c.ping(err => {
+      c.ping((err) => {
         if (err) {
           localErr.message = err.message;
           localErr.code = err.code;
@@ -147,7 +147,7 @@ class PromiseConnection extends EventEmitter {
         } else {
           const wrappedStatement = new PromisePreparedStatementInfo(
             statement,
-            promiseImpl
+            promiseImpl,
           );
           resolve(wrappedStatement);
         }
@@ -159,7 +159,7 @@ class PromiseConnection extends EventEmitter {
     const c = this.connection;
     const localErr = new Error();
     return new this.Promise((resolve, reject) => {
-      c.changeUser(options, err => {
+      c.changeUser(options, (err) => {
         if (err) {
           localErr.message = err.message;
           localErr.code = err.code;
@@ -192,14 +192,14 @@ class PromiseConnection extends EventEmitter {
     const func = functionsToWrap[i];
 
     if (
-      typeof Connection.prototype[func] === 'function' &&
+      typeof BaseConnection.prototype[func] === 'function' &&
       PromiseConnection.prototype[func] === undefined
     ) {
       PromiseConnection.prototype[func] = (function factory(funcName) {
         return function () {
-          return Connection.prototype[funcName].apply(
+          return BaseConnection.prototype[funcName].apply(
             this.connection,
-            arguments
+            arguments,
           );
         };
       })(func);
@@ -216,6 +216,7 @@ class PromiseConnection extends EventEmitter {
   'pause',
   'pipe',
   'resume',
-  'unprepare'
+  'unprepare',
 ]);
-module.exports = PromiseConnection
+
+module.exports = PromiseConnection;
