+++ lib/promise_pool.js
@@ -4,7 +4,7 @@ const EventEmitter = require('events').EventEmitter;
 const makeDoneCb = require('./make_done_cb.js');
 const PromisePoolConnection = require('./promise_pool_connection.js');
 const inheritEvents = require('./inherit_events.js');
-const Pool = require('./pool');
+const BasePool = require('./base_pool.js');
 
 class PromisePool extends EventEmitter {
   constructor(pool, thePromise) {
@@ -36,7 +36,7 @@ class PromisePool extends EventEmitter {
     const localErr = new Error();
     if (typeof args === 'function') {
       throw new Error(
-        'Callback function is not available with promise clients.'
+        'Callback function is not available with promise clients.',
       );
     }
     return new this.Promise((resolve, reject) => {
@@ -54,7 +54,7 @@ class PromisePool extends EventEmitter {
     const localErr = new Error();
     if (typeof args === 'function') {
       throw new Error(
-        'Callback function is not available with promise clients.'
+        'Callback function is not available with promise clients.',
       );
     }
     return new this.Promise((resolve, reject) => {
@@ -71,7 +71,7 @@ class PromisePool extends EventEmitter {
     const corePool = this.pool;
     const localErr = new Error();
     return new this.Promise((resolve, reject) => {
-      corePool.end(err => {
+      corePool.end((err) => {
         if (err) {
           localErr.message = err.message;
           localErr.code = err.code;
@@ -92,12 +92,12 @@ class PromisePool extends EventEmitter {
     const func = functionsToWrap[i];
 
     if (
-      typeof Pool.prototype[func] === 'function' &&
+      typeof BasePool.prototype[func] === 'function' &&
       PromisePool.prototype[func] === undefined
     ) {
       PromisePool.prototype[func] = (function factory(funcName) {
         return function () {
-          return Pool.prototype[funcName].apply(this.pool, arguments);
+          return BasePool.prototype[funcName].apply(this.pool, arguments);
         };
       })(func);
     }
@@ -106,7 +106,7 @@ class PromisePool extends EventEmitter {
   // synchronous functions
   'escape',
   'escapeId',
-  'format'
+  'format',
 ]);
 
 module.exports = PromisePool;
