+++ lib/promise_pool_connection.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const PromiseConnection = require('./promise_connection.js');
-const PoolConnection = require('./pool_connection.js');
+const BasePoolConnection = require('./base_pool_connection.js');
 
 class PromisePoolConnection extends PromiseConnection {
   constructor(connection, promiseImpl) {
@@ -9,9 +9,9 @@ class PromisePoolConnection extends PromiseConnection {
   }
 
   destroy() {
-    return PoolConnection.prototype.destroy.apply(
+    return BasePoolConnection.prototype.destroy.apply(
       this.connection,
-      arguments
+      arguments,
     );
   }
 }
