+++ lib/promise_prepared_statement_info.js
@@ -22,11 +22,11 @@ class PromisePreparedStatementInfo {
   }
 
   close() {
-    return new this.Promise(resolve => {
+    return new this.Promise((resolve) => {
       this.statement.close();
       resolve();
     });
   }
 }
 
-module.exports = PromisePreparedStatementInfo
+module.exports = PromisePreparedStatementInfo;
